+++ Controllers/item.js
@@ -1,10 +1,10 @@
-import ItemDetails from "../Models/itemdetails.js";
+import item from "../Models/item.js";
 
 // Get all items
 
 const getAllItems = async (req, res) => {
     try {
-        const items = await ItemDetails.find({});
+        const items = await item.find({});
         res.status(200).json(items);
     } catch (error) {
         res.status(500).send(error.message);
@@ -14,7 +14,7 @@ const getAllItems = async (req, res) => {
 // Get an item by ID
 const getItem = async (req, res) => {
     try {
-        const item = await ItemDetails.findById(req.params.id);
+        const item = await item.findById(req.params.id);
         if (!item) {
             return res.status(404).send('Item not found');
         }
@@ -27,7 +27,7 @@ const getItem = async (req, res) => {
 // Add a new item
 const addItem = async (req, res) => {
     try {
-        const newItem = new ItemDetails(req.body);
+        const newItem = new item(req.body);
         await newItem.save();
         res.status(201).json(newItem);
     } catch (error) {
@@ -38,7 +38,7 @@ const addItem = async (req, res) => {
 // Update an item
 const updateItem = async (req, res) => {
     try {
-        const updatedItem = await ItemDetails.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        const updatedItem = await item.findByIdAndUpdate(req.params.id, req.body, { new: true });
         if (!updatedItem) {
             return res.status(404).send('Item not found');
         }
@@ -51,7 +51,7 @@ const updateItem = async (req, res) => {
 // Delete an item
 const deleteItem = async (req, res) => {
     try {
-        const item = await ItemDetails.findByIdAndDelete(req.params.id);
+        const item = await item.findByIdAndDelete(req.params.id);
         if (!item) {
             return res.status(404).send('Item not found');
         }
