+++ Models/closet.js
@@ -11,12 +11,12 @@ const closetSchema = new mongoose.Schema({
         required: true, 
         ref: 'Profile'
     },
-    category: {
-        type: String,
-        required: true,
-        enum: ['outerwear', 'footwear', 'clothing', 'accessories'],
-    },
-    items: [{ type: Schema.Types.ObjectId, ref: 'Item' }]
+    // category: {
+    //     type: String,
+    //     required: true,
+    //     enum: ['outerwear', 'footwear', 'clothing', 'accessories'],
+    // },
+    item: [{ type: Schema.Types.ObjectId, ref: 'Item' }]
 });
 
 const Closet = mongoose.model('Closet', closetSchema);
